requirements.txt
pandas
psycopg2-binary
SQLAlchemy==1.4.46
streamlit
streamlit-authenticator
streamlit-aggrid
